requirements.txt
discord.py
chat-exporter
requests
